+++ crypto/addresses/Cargo.toml
@@ -18,6 +18,7 @@ workflow-wasm.workspace = true
 
 [dev-dependencies]
 criterion.workspace = true
+wasm-bindgen-test = "0.3.37"
 web-sys = "0.3.64"
 
 [[bench]]
